Debugging in Prod (During Demo)

+++ src/include/process.h
@@ -5,19 +5,19 @@
 #include <inttypes.h>
 
 typedef struct {
-    uint32_t pid;
-    uint32_t arrival_time;
-    uint32_t burst_time;
-    uint32_t remaining_burst;
-    uint32_t waiting_time;
-    uint32_t start_time[ARR_MAX];
-    uint32_t end_time[ARR_MAX];
+    uint64_t pid;
+    uint64_t arrival_time;
+    uint64_t burst_time;
+    uint64_t remaining_burst;
+    uint64_t waiting_time;
+    uint64_t start_time[ARR_MAX];
+    uint64_t end_time[ARR_MAX];
     size_t run_count;
 } Process;
 
-Process* createProcess(uint32_t pid, uint32_t arrival_time, uint32_t burst_time);
-void appendStartTime(Process *p, uint32_t start);
-void appendEndTime(Process *p, uint32_t end);
+Process* createProcess(uint64_t pid, uint64_t arrival_time, uint64_t burst_time);
+void appendStartTime(Process *p, uint64_t start);
+void appendEndTime(Process *p, uint64_t end);
 
 typedef struct {
     Process **processes;
